Fix the MPI datatype for DataPoint

+++ src/riversys.c
@@ -1711,7 +1711,7 @@ int Load_Forcings(
     // Create an MPI type for struct DataPoint
     const int nitems = 2;
     int          blocklengths[2] = { 1,1 };
-    MPI_Datatype types[2] = { MPI_INT, MPI_INT };
+    MPI_Datatype types[2] = { MPI_DOUBLE, MPI_FLOAT };
     MPI_Datatype mpi_datapoint_type;
     MPI_Aint     offsets[2];
